pyproject.toml
[project]
name = "ai-interview-coach"
version = "0.1.0"
description = "AI-powered Interview Coach built with LangGraph and OpenRouter"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=1.0.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-core>=0.3.0",
    "streamlit>=1.40.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/interview_coach"]

[project.scripts]
interview-coach = "interview_coach.graph:build_interview_graph"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "ruff>=0.8.0",
]
